Fix: constants 구조 변경
