+++ src/js/constants/errorMessage.js
@@ -1,0 +1,4 @@
+export const ERROR_MESSAGE = {
+  INVALID_LOTTO_PRICE_UNIT: '로또 구입 금액은 1,000원 단위로만 입력이 가능합니다.',
+  INVALID_NEGATIVE_LOTTO_PRICE: '로또 구입 금액에 음수는 입력할 수 없습니다.',
+};
