+++ src/js/constants/index.js
@@ -1,0 +1,3 @@
+export * from './lotto.js';
+export * from './selector.js';
+export * from './errorMessage.js';
