+++ src/js/constants/lotto.js
@@ -1,21 +1,6 @@
-export const LOTTO_PRICE = 1_000;
-
-export const LOTTO_START_NUMBER = 1;
-export const LOTTO_END_NUMBER = 45;
-export const LOTTO_NUMBER_COUNT_LIMIT = 6;
-
-export const SELECTOR = {
-  PAYMENT_FORM: '.js-payment-form',
-  PURCHASE_PRICE_INPUT: '.js-purchase-price-input',
-  PURCHASED_LOTTO: '.js-purchased-lotto',
-  LOTTO_PURCHASE_COUNT_TEXT: '.js-lotto-purchase-count-text',
-  LOTTO_DETAIL_NUMBER: '.js-lotto-detail-number',
-  LOTTO_NUMBERS_TOGGLE_BUTTON: '.js-lotto-numbers-toggle-button',
-  LOTTO_ICON_WRAPPER: '.js-lotto-icon-wrapper',
-  LOTTO_RESULT_FORM: '.js-lotto-result-form',
-};
-
-export const ERROR_MESSAGE = {
-  INVALID_LOTTO_PRICE_UNIT: '로또 구입 금액은 1,000원 단위로만 입력이 가능합니다.',
-  INVALID_NEGATIVE_LOTTO_PRICE: '로또 구입 금액에 음수는 입력할 수 없습니다.',
+export const LOTTO = {
+  PRICE: 1_000,
+  MIN_NUMBER: 1,
+  MAX_NUMBER: 45,
+  NUMBER_AMOUNT: 6,
 };
