+++ src/js/constants/selector.js
@@ -1,0 +1,10 @@
+export const SELECTOR = {
+  PAYMENT_FORM: '.js-payment-form',
+  PURCHASE_PRICE_INPUT: '.js-purchase-price-input',
+  PURCHASED_LOTTO: '.js-purchased-lotto',
+  LOTTO_PURCHASE_COUNT_TEXT: '.js-lotto-purchase-count-text',
+  LOTTO_DETAIL_NUMBER: '.js-lotto-detail-number',
+  LOTTO_NUMBERS_TOGGLE_BUTTON: '.js-lotto-numbers-toggle-button',
+  LOTTO_ICON_WRAPPER: '.js-lotto-icon-wrapper',
+  LOTTO_RESULT_FORM: '.js-lotto-result-form',
+};
