+++ src/js/eventHandlers/lotto.js
@@ -1,4 +1,4 @@
-import { ERROR_MESSAGE, LOTTO_PRICE } from '../constants/lotto.js';
+import { ERROR_MESSAGE, LOTTO } from '../constants/index.js';
 import { validation } from '../utils/validation.js';
 import {
   getLottoPurchasePrice,
@@ -18,11 +18,11 @@ export const handleSubmit = (e) => {
     if (!validation.isPositiveNumber(lottoPurchasePrice)) {
       throw Error(ERROR_MESSAGE.INVALID_NEGATIVE_LOTTO_PRICE);
     }
-    if (!validation.isRemainderZero(lottoPurchasePrice, LOTTO_PRICE)) {
+    if (!validation.isRemainderZero(lottoPurchasePrice, LOTTO.PRICE)) {
       throw Error(ERROR_MESSAGE.INVALID_LOTTO_PRICE_UNIT);
     }
 
-    const lottoCount = lottoPurchasePrice / LOTTO_PRICE;
+    const lottoCount = lottoPurchasePrice / LOTTO.PRICE;
 
     renderLottoPurchaseCountText(lottoCount);
     renderLottoIcons(lottoCount);
